feat: add QR code to share dialog update about dialog to include company info
