+++ lib/assets.dart
@@ -4,4 +4,6 @@ class AppAssets {
 
   static const String PNG_LOGO_DARK = 'assets/png/wvems_logo_dark.png';
   static const String PNG_LOGO_LIGHT = 'assets/png/wvems_logo_light.png';
+
+  static const String PNG_QR_CODE = 'assets/png/share_qr_code.png';
 }
