+++ lib/ui/views/nav_dialogs/dialogs.dart
@@ -1,4 +1,4 @@
-export 'about_dialog.dart';
+export 'about_dialog/about_dialog.dart';
 export 'messages_dialog.dart';
 export 'settings_dialog/settings_dialog.dart';
-export 'share_dialog.dart';
+export 'share_dialog/share_dialog.dart';
