+++ lib/ui/views/nav_dialogs/shared/nav_divider.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class NavDivider extends StatelessWidget {
+  const NavDivider({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Divider(color: Theme.of(context).accentColor, thickness: 1);
+  }
+}
